+++ app/src/main/java/de/tum/in/tumcampusapp/component/tumui/feedback/FeedbackContract.kt
@@ -25,11 +25,11 @@ interface FeedbackContract {
         fun showProgressDialog()
         fun showSendErrorDialog()
         fun onFeedbackSent()
-        fun openCamera(intent: Intent)
-        fun openGallery(intent: Intent)
-        fun showPermissionRequestDialog(permission: String, requestCode: Int)
-        fun onImageAdded(path: String)
-        fun onImageRemoved(position: Int)
+       // fun openCamera(intent: Intent)
+        //fun openGallery(intent: Intent)
+        //fun showPermissionRequestDialog(permission: String, requestCode: Int)
+        //fun onImageAdded(path: String)
+        //fun onImageRemoved(position: Int)
     }
 
     interface Presenter {
@@ -38,14 +38,14 @@ interface FeedbackContract {
         fun attachView(view: View)
         fun onRestoreInstanceState(savedInstanceState: Bundle)
         fun initEmail()
-        fun removeImage(path: String)
+        //fun removeImage(path: String)
         fun onSendFeedback()
         fun onConfirmSend()
-        fun onImageOptionSelected(option: Int)
-        fun onNewImageTaken()
-        fun onNewImageSelected(uri: Uri?)
-        fun takePicture()
-        fun openGallery()
+        //fun onImageOptionSelected(option: Int)
+        //fun onNewImageTaken()
+        //fun onNewImageSelected(uri: Uri?)
+        //fun takePicture()
+        //fun openGallery()
         fun listenForLocation()
         fun onSaveInstanceState(outState: Bundle)
         fun detachView()
