+++ app/src/main/java/de/tum/in/tumcampusapp/utils/camera/CameraUtils.kt
@@ -1,0 +1,4 @@
+package de.tum.`in`.tumcampusapp.utils.camera
+
+class CameraUtils {
+}
